SAK-67 separeted inStock component
